+++ Backend/modif.sql
@@ -1,0 +1,5 @@
+DELETE FROM `works`
+WHERE `id` = 31
+
+DELETE FROM `works`
+WHERE `id` = 32
