+++ FrontEnd/login.html
@@ -3,13 +3,13 @@
 <head>
 	<meta charset="UTF-8">
 	<title>Sophie Bluel - Architecte d'intérieur</title> 
-	<script src="./assets/login.js" type="module" defer></script>
+	<script src="./JS/login.js" type="module" defer></script>
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<link rel="preconnect" href="https://fonts.googleapis.com">
 	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 	<link href="https://fonts.googleapis.com/css2?family=Syne:wght@700;800&family=Work+Sans&display=swap" rel="stylesheet">
 	<meta name="description" content="">
-	<link rel="stylesheet" href="./assets/style.css">
+	<link rel="stylesheet" href="./CSS/style.css">
 </head>
 <body>
 	<div class="main-container">
@@ -29,9 +29,9 @@
 			<section id="login-section">     
 				<form action="#" method="post">
 					<label for="email">Email</label>
-					<input type="email" name="email" id="email" required>
+					<input value="sophie.bluel@test.tld" type="email" name="email" id="email" required>
 					<label for="message">Mot de passe</label>
-					<input type="password" name="password" id="password" required>
+					<input value="S0phie" type="password" name="password" id="password" required>
 					<input type="submit" value="Se connecter">
 				</form>
 				<a href="#"> Mot de passe oublié</a>
